fix: Move talon files and docs to a subfolder
The dev shell environment triggered from the flake.nix file will create
a .direnv file, which contains a significant number of files/folders.
Talon doesn't seem to handle this well as it's constantly parsing it
looking for files to process (confirmed via strace).

The solution for now is to move everything into a subfolder (vim) which
I then link directly to from ~/.talon/user/ . For people not using nix,
it should be fine to clone the folder directly into ~/.talon/user/ if
preferred.