refactor: change app detection and mode behavior
Move towards using app-based matching for terminal apps. We also start
to differentiate between MODE:t and MODE:nt. I mark this with TERM-N: in
the vim title for now, but this may change. This requires an autocmd
change in the VIM config.

This allows us to have commands that only work when the application is
active and in terminal mode (for example backtrace via app: gdb), but
other commands specific to the application that also work when in a
terminal mode normal buffer, (for example go last prompt via tag: gdb).

The foo-nt.talon files for setting the tag based on TERM-N should likely
change to something less cluttery if possible.

Note these changes also do require some modifications to community
scripts, as previously files were relying only on tag: matching, and now
will differentiate.

There's another problem here to for some things like how do we
differentiate and active plugin (ex: pwndbg while in MODE:t) vs an
inactive, but we still maybe want some commands to work in MODE:nt. the
app is gdb, so we can't use app vs tag. We must have two tags, like
user.pwndbg and user.pwndbg-inactive. This is something that needs to be
worked out still.